refactors streams into video liveview
refactors streams into video liveview
