+++ lib/glimesh_web/live/chat_live/index.ex
@@ -16,12 +16,13 @@ defmodule GlimeshWeb.ChatLive.Index do
       if session["locale"], do: Gettext.put_locale(session["locale"])
       if connected?(socket), do: Streams.subscribe_to(:chat, channel_id)
 
+      user = Accounts.get_user_by_session_token(session["user_token"])
       channel = ChannelLookups.get_channel!(channel_id)
 
       # Sets a default user_preferences map for the chat if the user is logged out
       user_preferences =
-        if session["user"] do
-          Accounts.get_user_preference!(session["user"])
+        if user do
+          Accounts.get_user_preference!(user)
         else
           %{
             show_timestamps: false,
@@ -29,9 +30,7 @@ defmodule GlimeshWeb.ChatLive.Index do
           }
         end
 
-      if session["user"] do
-        user = session["user"]
-
+      if user do
         Presence.track_presence(
           self(),
           Streams.get_subscribe_topic(:chatters, channel.id),
@@ -51,7 +50,7 @@ defmodule GlimeshWeb.ChatLive.Index do
         |> assign(:channel_chat_parser_config, Chat.get_chat_parser_config(channel))
         |> assign(:update_action, "replace")
         |> assign(:channel, channel)
-        |> assign(:user, session["user"])
+        |> assign(:user, user)
         |> assign(:theme, Map.get(session, "site_theme", "dark"))
         |> assign(:permissions, Chat.get_moderator_permissions(channel, session["user"]))
         |> assign(:chat_messages, list_chat_messages(channel))
