+++ lib/glimesh_web/live/streams_live/video.html.leex
@@ -1,0 +1,149 @@
+<div class="card">
+  <div class="card-header p-1">
+    <div class="row">
+      <div class="col align-self-center d-block text-truncate">
+        <%= live_render(@socket, GlimeshWeb.UserLive.Components.ChannelTitle, id: "channel-title", session: %{"user" => @user, "channel_id" => @channel.id}) %>
+      </div>
+      <div class="col-auto">
+        <div class="float-right">
+          <div class="stream-info btn-toolbar">
+            <div class="btn-group mr-1" role="group" aria-label="First group">
+              <%= if @can_receive_payments do %>
+                <%= live_render(@socket, GlimeshWeb.UserLive.Components.SubscribeButton, id: "subscribe-button", session: %{"user" => @user, "streamer" => @streamer}) %>
+              <% end %>
+            </div>
+            <div class="btn-group mr-1" role="group" aria-label="Second group">
+              <%= live_render(@socket, GlimeshWeb.UserLive.Components.FollowButton, id: "follow-button", session: %{"user" => @user, "streamer" => @streamer}) %>
+            </div>
+            <div class="btn-group" role="group" aria-label="Third group">
+              <%= live_render(@socket, GlimeshWeb.UserLive.Components.ViewerCount, id: "viewer-count", session: %{"channel_id" => @channel.id}) %>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+  <div class="card-body p-0 p-md-1">
+    <%= if @prompt_mature do %>
+      <div class="jumbotron jumbotron-fluid jumbotron-mature-content mb-0">
+        <div class="container p-5">
+          <div class="row justify-content-center">
+            <div class="col-sm-6">
+              <h3 class="display-5 text-center"><%= gettext("Mature Content Warning") %></h3>
+              <p class="lead text-center">
+              <%= gettext("The streamer has flagged this channel as only appropriate for Mature Audiences.") %><br>
+              <%= gettext("Do you wish to continue?") %>
+              </p>
+            </div>
+            <div class="w-100"></div>
+            <div class="col-8 col-lg-3">
+              <button type="button" phx-click="show_mature" class="btn btn-primary btn-block"><%= gettext("Agree & View Channel") %></button>
+            </div>
+          </div>
+
+        </div>
+      </div>
+
+    <% else %>
+      <div id="video-container" class="embed-responsive embed-responsive-16by9">
+        <video id="video-player" class="embed-responsive-item" phx-hook="FtlVideo" controls playsinline poster="<%= @channel_poster %>"></video>
+        <div id="video-loading-container" class="">
+          <div class="lds-ring">
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+          </div>
+        </div>
+      </div>
+    <% end %>
+  </div>
+  <div class="card-footer p-1 d-none d-sm-block">
+    <div class="row">
+      <div class="col-8 d-inline-flex align-items-center">
+        <div id="streamer-avatar">
+          <a href="<%= Routes.user_profile_path(@socket, :index, @channel.user.username) %>">
+            <img src="<%= Glimesh.Avatar.url({@channel.user.avatar, @channel.user}, :original)%>" alt="<%= @channel.user.displayname %>" width="48" height="48" class="img-avatar mr-2 float-left <%= if Glimesh.Accounts.can_receive_payments?(@channel.user), do: "img-verified-streamer", else: "" %>">
+          </a>
+        </div>
+        <a title="<%= gettext("View Profile") %>" class="<%= Glimesh.Chat.Effects.get_username_color(@channel.user) %>" href="<%= Routes.user_profile_path(@socket, :index, @channel.user.username) %>">
+          <h3 class="mb-0"><%= @channel.user.displayname %></h3>
+        </a>
+
+        <span class="badge badge-pill badge-info ml-2"><%= Glimesh.Streams.get_channel_language(@channel) %></span>
+        <%= if @channel.mature_content do %>
+          <span class="badge badge-pill badge-warning ml-2"><%= gettext("Mature") %></span>
+        <% end %>
+
+        <%= live_render(@socket, GlimeshWeb.UserLive.Components.SocialButtons, id: "social-buttons", container: {:ul, class: "list-inline ml-2 mb-0"}, session: %{"user_id" => @streamer.id}) %>
+      </div>
+      <div class="col-4 align-self-center">
+        <div class="float-right mr-sm-2">
+          <div class="d-inline-block mr-sm-2">
+            <a href="#" phx-click="toggle_debug" class="text-color-link <%= if @player_error, do: "text-warning", else: "" %>">
+              <i class="fas fa-signal"></i>
+              <span class="sr-only">Debug</span>
+            </a>
+          </div>
+          <div class="d-inline-block">
+            <%= live_render(@socket, GlimeshWeb.UserLive.Components.ReportButton, id: "report-button", session: %{"user" => @user, "streamer" => @streamer}) %>
+          </div>
+        </div>
+      </div>
+    </div>
+    <%= if @show_debug do %>
+      <div id="debugModal" class="live-modal" phx-capture-click="toggle_debug" phx-window-keydown="toggle_debug" phx-key="escape" phx-target="#debugModal" phx-page-loading>
+        <div class="modal-dialog" role="document">
+          <div class="modal-content">
+            <div class="modal-header">
+              <h5 class="modal-title"><%= gettext("Debug Information") %></h5>
+              <button type="button" class="close" phx-click="toggle_debug" aria-label="Close">
+                <span aria-hidden="true">&times;</span>
+              </button>
+            </div>
+
+            <div class="modal-body">
+              <%= if @player_error do %>
+                <div class="alert alert-warning" role="alert">
+                  <%= @player_error %>
+                </div>
+              <% end %>
+              <pre class="px-2">
+== Janus Edge Information ==
+Edge Hostname: <%= @janus_hostname %>
+Edge URL: <%= @janus_url %>
+Reported Lost Packets: <%= @lost_packets %>
+<br>
+== Stream Metadata ==
+# Stream Metadata no longer auto reloads,
+# close & reopen modal to see newest data.
+ingest_server: <%= @stream_metadata.ingest_server %>
+ingest_viewers: <%= @stream_metadata.ingest_viewers %> // unused
+stream_time_seconds: <%= @stream_metadata.stream_time_seconds %>
+
+source_bitrate: <%= @stream_metadata.source_bitrate %>
+source_ping: <%= @stream_metadata.source_ping %> // unused
+
+recv_packets: <%= @stream_metadata.recv_packets %>
+lost_packets: <%= @stream_metadata.lost_packets %> // unused
+nack_packets: <%= @stream_metadata.nack_packets %> // unused
+
+vendor_name: <%= @stream_metadata.vendor_name %>
+vendor_version: <%= @stream_metadata.vendor_version %>
+
+audio_codec: <%= @stream_metadata.audio_codec %>
+video_codec: <%= @stream_metadata.video_codec %>
+video_height: <%= @stream_metadata.video_height %> // unused
+video_width: <%= @stream_metadata.video_width %> // unused
+
+inserted_at: <%= @stream_metadata.inserted_at %>
+updated_at: <%= @stream_metadata.updated_at %>
+              </pre>
+            </div>
+
+          </div>
+        </div>
+      </div>
+    <% end %>
+  </div>
+</div>
