+++ priv/repo/mocks/streams.exs
@@ -4,8 +4,6 @@ possible_categories = Enum.map(Glimesh.ChannelCategories.list_categories(), fn x
 
 for n <- 1..num_streams do
   category_id = Enum.random(possible_categories)
-  possible_tags = Glimesh.ChannelCategories.list_tags(category_id)
-  tags = Enum.take_random(possible_tags, :rand.uniform(length(possible_tags)))
 
   {:ok, user} =
     Glimesh.Accounts.register_user(%{
@@ -22,8 +20,6 @@ for n <- 1..num_streams do
 
   {:ok, channel} =
     channel
-    |> Glimesh.Repo.preload(:tags)
     |> Ecto.Changeset.change()
-    |> Ecto.Changeset.put_assoc(:tags, tags)
     |> Glimesh.Repo.update()
 end
